+++ LEDGridPatterns.h
@@ -1,0 +1,8 @@
+#ifndef LEDGridPatterns_H
+#define LEDGridPatterns_H
+#include <cstdint>
+#include <FastLED.h>
+
+void showPulsatingBars(uint8_t hue, CRGB* leds);
+
+#endif
